+++ docs/init.js
@@ -1,4 +1,0 @@
-window.kempo = {
-	pathToStylesheet: './kempo.min.css',
-	pathsToIcons: ['https://cdn.jsdelivr.net/npm/kempo-ui@0.1.6/icons']
-};